+++ actix-session/CHANGES.md
@@ -1,11 +1,11 @@
 # Changes
 
-## unreleased
+## Unreleased - 2020-xx-xx
+* Update `actix-web` dependency to 3.0.0.
+* Minimum supported Rust version(MSRV) is now 1.42.0.
 
-* Minimum supported Rust version(MSRV) is now 1.40.0.
 
 ## [0.4.0-alpha.1] - 2020-03-14
-
 * Update the `time` dependency to 0.2.7
 * Update the `actix-web` dependency to 3.0.0-alpha.1
 * Long lasting auto-prolonged session [#1292]